Can now do things like "sag -wp"
Install script now doesn't overwrite existing .sackrc. Optimized awk
loop in-case thousands of matches are found. Now if sack__cwd is
empty, use pwd.

+++ install_sack.sh
@@ -25,4 +25,6 @@ cp $sack__install_cwd/sack $sack__install_main/
 chmod +x $sack__install_main/sack
 cp $sack__install_cwd/sag $sack__install_main/
 chmod +x $sack__install_main/sag
-cp $sack__install_cwd/.sackrc $sack__install_sackrc/
+if [[ ! -e $sack__install_sackrc/.sackrc ]]; then
+    cp $sack__install_cwd/.sackrc $sack__install_sackrc/
+fi
